multi_tool_agent/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
aiosqlite==0.22.0
alembic==1.17.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
appdirs==1.4.4
arxiv==2.3.1
asgiref==3.11.0
asttokens==3.0.1
attrs==25.4.0
Authlib==1.6.6
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.14.3
blinker==1.9.0
brotli==1.2.0
build==1.3.0
cachetools==6.2.3
certifi==2025.11.12
cffi==2.0.0
cfgv==3.5.0
charset-normalizer==3.4.4
chroma-hnswlib==0.7.6
chromadb==0.5.23
click==8.3.1
cloudpickle==3.1.2
cohere==5.20.0
coloredlogs==15.0.1
crewai==0.140.0
crewai-tools==0.49.0
cryptography==46.0.3
curl_cffi==0.13.0
dataclasses-json==0.6.7
ddgs==9.9.3
decorator==5.2.1
deprecation==2.1.0
diskcache==5.6.3
distlib==0.4.0
distro==1.9.0
docker==7.1.0
docstring_parser==0.17.0
durationpy==0.10
embedchain==0.1.128
et_xmlfile==2.0.0
executing==2.2.1
fake-useragent==2.2.0
fastapi==0.123.10
fastavro==1.12.1
feedparser==6.0.12
filelock==3.20.0
flatbuffers==25.9.23
frozendict==2.4.7
frozenlist==1.8.0
fsspec==2025.12.0
google-adk==1.21.0
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-cloud-aiplatform==1.130.0
google-cloud-appengine-logging==1.7.0
google-cloud-audit-log==0.4.0
google-cloud-bigquery==3.38.0
google-cloud-bigquery-storage==2.35.0
google-cloud-bigtable==2.34.0
google-cloud-core==2.5.0
google-cloud-discoveryengine==0.13.12
google-cloud-logging==3.12.1
google-cloud-monitoring==2.28.0
google-cloud-resource-manager==1.15.0
google-cloud-secret-manager==2.25.0
google-cloud-spanner==3.60.0
google-cloud-speech==2.34.0
google-cloud-storage==3.7.0
google-cloud-trace==1.17.0
google-crc32c==1.7.1
google-genai==1.55.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
gptcache==0.1.44
graphviz==0.21
grpc-google-iam-v1==0.14.3
grpc-interceptor==0.15.4
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.0
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.36.0
humanfriendly==10.0
hyperframe==6.1.0
identify==2.6.15
idna==3.11
importlib_metadata==8.7.0
importlib_resources==6.5.2
instructor==1.12.0
ipython==9.8.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
json5==0.12.1
json_repair==0.25.2
jsonpatch==1.33
jsonpickle==4.1.1
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kubernetes==34.1.0
lance-namespace==0.3.1
lance-namespace-urllib3-client==0.3.1
lancedb==0.25.3
langchain==0.3.27
langchain-cohere==0.3.5
langchain-community==0.3.31
langchain-core==0.3.80
langchain-experimental==0.3.4
langchain-openai==0.2.14
langchain-text-splitters==0.3.11
langsmith==0.3.45
litellm==1.72.6
lxml==6.0.2
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib-inline==0.2.1
mcp==1.24.0
mdurl==0.1.2
mem0ai==0.1.118
mmh3==5.2.0
monotonic==1.6
mpmath==1.3.0
multidict==6.7.0
multitasking==0.0.12
mypy_extensions==1.1.0
networkx==3.6.1
nodeenv==1.9.1
numpy==2.3.5
oauthlib==3.3.1
onnxruntime==1.22.0
openai==1.109.1
openpyxl==3.1.5
opentelemetry-api==1.37.0
opentelemetry-exporter-gcp-logging==1.11.0a0
opentelemetry-exporter-gcp-monitoring==1.11.0a0
opentelemetry-exporter-gcp-trace==1.11.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-exporter-otlp-proto-http==1.37.0
opentelemetry-instrumentation==0.60b1
opentelemetry-instrumentation-asgi==0.60b1
opentelemetry-instrumentation-fastapi==0.60b1
opentelemetry-proto==1.37.0
opentelemetry-resourcedetector-gcp==1.11.0a0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
opentelemetry-util-http==0.60b1
orjson==3.11.5
overrides==7.7.0
packaging==25.0
pandas==2.3.3
parso==0.8.5
pdfminer.six==20251107
pdfplumber==0.11.8
peewee==3.18.3
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.1
portalocker==3.2.0
posthog==3.25.0
pre_commit==4.5.0
primp==0.15.0
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.5
pypdf==5.9.0
pypdfium2==5.2.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyright==1.1.407
pysbd==0.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytube==15.0.0
pytz==2025.2
pyvis==0.3.2
PyYAML==6.0.3
qdrant-client==1.16.2
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.30.0
rsa==4.9.1
schema==0.7.8
sgmllib3k==1.0.0
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.8
SQLAlchemy==2.0.45
sqlalchemy-spanner==1.17.1
sqlparse==0.5.4
sse-starlette==3.0.3
stack-data==0.6.3
starlette==0.50.0
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
tiktoken==0.12.0
tokenizers==0.20.3
tomli==2.3.0
tomli_w==1.2.0
tqdm==4.67.1
traitlets==5.14.3
typer==0.20.0
types-requests==2.32.4.20250913
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.3.0
uv==0.9.17
uvicorn==0.38.0
uvloop==0.22.1
virtualenv==20.35.4
watchdog==6.0.0
watchfiles==1.1.1
wcwidth==0.2.14
websocket-client==1.9.0
websockets==15.0.1
wrapt==1.17.3
yarl==1.22.0
yfinance==0.2.64
zipp==3.23.0
zstandard==0.23.0
